+++ src/dfx-core/src/error/canister.rs
@@ -4,4 +4,16 @@ use thiserror::Error;
 pub enum CanisterBuilderError {
     #[error("Failed to construct wallet canister caller: {0}")]
     WalletCanisterCaller(ic_agent::AgentError),
+
+    #[error("Failed to build call sender: {0}")]
+    CallSenderBuildError(ic_agent::AgentError),
+}
+
+#[derive(Error, Debug)]
+pub enum CanisterInstallError {
+    #[error(transparent)]
+    CanisterBuilderError(#[from] CanisterBuilderError),
+
+    #[error("Failed during wasm installation call: {0}")]
+    InstallWasmError(ic_agent::AgentError),
 }
