+++ src/dfx/src/commands/canister/install.rs
@@ -1,9 +1,10 @@
 use crate::lib::canister_info::CanisterInfo;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
-use crate::lib::operations::canister::{install_canister, install_canister_wasm};
+use crate::lib::operations::canister::install_canister;
 use crate::lib::root_key::fetch_root_key_if_needed;
 use crate::util::{blob_from_arguments, get_candid_init_type};
+use dfx_core::canister::install_canister_wasm;
 use dfx_core::identity::CallSender;
 
 use anyhow::{anyhow, bail, Context};
@@ -113,7 +114,6 @@ pub async fn exec(
             let mode = mode.context("The install mode cannot be auto when using --wasm")?;
             let install_args = blob_from_arguments(arguments, None, arg_type, &None)?;
             install_canister_wasm(
-                env,
                 agent,
                 canister_id,
                 canister_info.as_ref().map(|info| info.get_name()).ok(),
@@ -123,8 +123,10 @@ pub async fn exec(
                 fs::read(&wasm_path)
                     .with_context(|| format!("Unable to read {}", wasm_path.display()))?,
                 opts.yes,
+                env.get_logger(),
             )
             .await
+            .map_err(Into::into)
         } else {
             let canister_info = canister_info
                 .with_context(|| format!("Failed to load canister info for {}.", canister))?;
@@ -150,6 +152,7 @@ pub async fn exec(
                 !opts.no_asset_upgrade,
             )
             .await
+            .map_err(Into::into)
         }
     } else if opts.all {
         // Install all canisters.
