+++ src/dfx/src/lib/operations/canister/mod.rs
@@ -4,9 +4,10 @@ mod install_canister;
 
 pub use create_canister::create_canister;
 pub use deploy_canisters::deploy_canisters;
+pub use dfx_core::canister::install_canister_wasm;
 use fn_error_context::context;
 use ic_utils::Argument;
-pub use install_canister::{install_canister, install_canister_wasm, install_wallet};
+pub use install_canister::{install_canister, install_wallet};
 
 use crate::lib::canister_info::CanisterInfo;
 use crate::lib::environment::Environment;
